requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
marshmallow==3.12.2
uuid==1.30
Flask-Migrate==3.1.0
python-dotenv==0.20.0
Flask-Script==2.0.6